Port influx tests
