+++ desktop/panel/database.test.js
@@ -73,76 +73,4 @@ for (const subprocess of RUNNERS) {
       }
     );
   }
-
-  describe('influx testdata/influx tests', () => {
-    const tests = [
-      {
-        address: 'localhost:8087',
-        query: 'SELECT MEAN(avg_wave_period_sec) FROM ndbc',
-        version: 'influx',
-      },
-      {
-        address: 'localhost:8086',
-        query: `
-	 from(bucket: "test")
-	 |> range(start: -1000000h)
-	 |> filter(fn: (r) =>
-           (r._measurement == "ndbc" and r._field == "avg_wave_period_sec"))
-	 |> group(columns: ["_measurement", "_start", "_stop", "_field"], mode: "by")
-	 |> keep(columns: ["_measurement", "_start", "_stop", "_field", "_time", "_value"])
-	 |> mean()
-	 |> map(fn: (r) =>
-           ({r with _time: 1970-01-01T00:00:00Z}))
-	 |> rename(columns: {_value: "mean", "_time": "time"})
-         |> drop(columns: ["result", "table"])
-	 |> yield(name: "0")`,
-        version: 'influx-flux',
-      },
-    ];
-
-    for (const testcase of tests) {
-      test(`runs ${JSON.stringify(testcase)} query`, async () => {
-        if (process.platform !== 'linux') {
-          return;
-        }
-
-        const connectors = [
-          new DatabaseConnectorInfo({
-            type: testcase.version,
-            address: testcase.address,
-            database: 'test',
-            username: 'test',
-            password_encrypt: new Encrypt('testtest'),
-            apiKey_encrypt: new Encrypt('test'),
-          }),
-        ];
-        const dp = new DatabasePanelInfo();
-        dp.database.connectorId = connectors[0].id;
-        dp.content = testcase.query;
-
-        let finished = false;
-        const panels = [dp];
-        await withSavedPanels(
-          panels,
-          async (project) => {
-            const panelValueBuffer = fs.readFileSync(
-              getProjectResultsFile(project.projectName) + dp.id
-            );
-
-            const v = JSON.parse(panelValueBuffer.toString());
-            expect(v.length).toEqual(1);
-            expect(v[0].time).toStrictEqual('1970-01-01T00:00:00Z');
-            expect(v[0].mean).toStrictEqual(6.6);
-
-            finished = true;
-          },
-          { evalPanels: true, connectors, subprocessName: subprocess }
-        );
-
-        if (!finished) {
-          throw new Error('Callback did not finish');
-        }
-      });
-    }
-  });
 }
